Utenti/App/requirements.txt
flask
mysql-connector-python
#import delle librerie